Refactor muscic album entry methods
