+++ data/music_album.json
@@ -2,3 +2,4 @@
 {"title":"we are great","publish_date":"1998-12-01","archived":false,"genre":"gpa"}
 {"title":"rap","publish_date":"2022-06-06","archived":false,"genre":"unknown"}
 {"title":"Tega","publish_date":"2012-05-06","archived":true,"genre":"jams"}
+{"title":"degs","publish_date":"2000-08-06","archived":true,"genre":"rats"}
