+++ main.rb
@@ -1,8 +1,6 @@
-require_relative 'class/item'
-require_relative 'class/music_album'
-require_relative 'class/genre'
 require_relative 'apps/game_app'
 require_relative 'apps/movie_app'
+require_relative 'apps/music_album_app'
 require 'json'
 
 
@@ -22,14 +20,14 @@ def display_menu
   puts '10 - Add a music album'
   puts '11 - Add a movie'
   puts '12 - Add a game'
-  puts '13 - Add a genre'
-  puts '14 - Add a label'
-  puts '15 - Add a author'
-  puts '16 - Exit'
+  puts '13 - Add a label'
+  puts '14 - Add a author'
+  puts '15 - Exit'
 end
 
 
 movie_app = MovieApp.new
+music_album_app = Music_albumApp.new
 app = GameApp.new
 loop do
   display_menu
@@ -39,23 +37,15 @@ loop do
     # your code here
   when 2
     # List All Music Albums
-    puts 'List of Music_albums:'
-    music_albums.each do |item|
-      puts "- Title: #{item.title}, Genre: #{item.genre}, Published: #{item.publish_date}, Archived: #{item.archived}"
-    end
+    music_album_app.list_music_albums
   when 3
     # List All Movies
-    #movie_app.movies_load_data
     movie_app.list_movies
   when 4
     app.list_games
   when 5
     # List All Genre
-    puts 'List of Genre:'
-    genres = music_albums.map(&:genre).uniq.compact
-    genres.each do |genre|
-      puts "- Name: #{genre}"
-    end
+    music_album_app.list_music_album_genres
   when 6
     # your code here
   when 7
@@ -67,44 +57,19 @@ loop do
     # your code here
   when 10
     # Add A Music
-    puts 'Enter music_album title:'
-    title = gets.chomp
-    puts 'Enter publish date (YYYY-MM-DD):'
-    publish_date = gets.chomp
-    puts 'Is it on_spotify? (y/n)'
-    on_spotify = true if gets.chomp.downcase == 'y'
-
-    music_album = MusicAlbum.new(title, publish_date, on_spotify: on_spotify)
-    music_album.move_to_archive
-    music_albums << music_album
-
-    puts 'Music_album added!'
+    music_album_app.add_music_album
   when 11
     # Add A Movie
    movie_app.add_movie
   when 12
     app.add_game
   when 13
-    # Add Music_album genre
-    puts 'Enter genre name:'
-    genre_name = gets.chomp
-
-    genre = Genre.new(genre_name)
-    music_albums.each do |item|
-      item.genre = genre.name if item.genre.nil?
-    end
-
-    puts 'Genre added!'
-  when 14
     # your code here
-  when 15
+  when 14
     app.add_author
-  when 16
-    # Save Movies data to JSON files before exiting
-    # movie_app.movies_save_data(movies)
-    # Save Music_album data to JSON files before exiting
-    # music_save_data(music_albums)
-    puts 'Thanks for using Catalog of My Things app!'
+  when 15
+    # Exit Console App
+    puts 'Thanks for using Personal Inventory app!'
     break
   else
     puts 'Invalid Option, please choose from 1-10'
